+++ ceammc/ext/src/synth/synth_piano.cpp
@@ -6,13 +6,11 @@
 using namespace ceammc;
 
 class SynthPiano : public SynthWithFreq<faust_synth_piano_tilde> {
-    UIProperty* gain_;
     ClockLambdaFunction clock_;
 
 public:
     SynthPiano(const PdArgs& args)
         : SynthWithFreq<faust_synth_piano_tilde>(args)
-        , gain_(static_cast<UIProperty*>(property(PROP_GAIN)))
         , clock_([this]() { gate_->setValue(0); })
     {
     }
@@ -20,39 +18,14 @@ public:
     void onBang() override
     {
         gate_->setValue(1);
+        dsp_->instanceClear();
         clock_.delay(150);
     }
 
-    void m_note(t_symbol* s, const AtomListView& lv)
+    void m_reset(t_symbol*, const AtomListView&)
     {
-        if (!pitch_ || !gate_ || !gain_) {
-            METHOD_ERR(s) << "dev error: @pitch or @gate or @gain property not found";
-            return;
-        }
-
-        if (lv.size() != 2 && !lv.allOf(isFloat)) {
-            METHOD_ERR(s) << "usage: " << s->s_name << " NOTE VELOCITY";
-            return;
-        }
-
-        const auto note = lv.floatAt(0, -1);
-        const auto vel = lv.floatAt(1, -1);
-        const auto nmin = pitch_->infoT().minFloat();
-        const auto nmax = pitch_->infoT().maxFloat();
-
-        if (note < nmin || note > nmax) {
-            METHOD_ERR(s) << "note value is out of [" << nmin << ".." << nmax << "] range: " << note;
-            return;
-        }
-
-        if (vel < 0 || vel > 127) {
-            METHOD_ERR(s) << "velocity value is out of [0..127] range: " << vel;
-            return;
-        }
-
-        pitch_->setValue(note, true);
-        gain_->setValue(vel / 127, true);
-        gate_->setValue(vel > 0, true);
+        gate_->setValue(0);
+        dsp_->instanceClear();
     }
 };