+++ ceammc/CHANGELOG.md
@@ -28,6 +28,7 @@
   - route.float (with route.f alias) added
   - system.exit added
   - synth.harpsichord~ added
+  - synth.piano~ added
   - ui.button added (with ui.btn alias)
   - ui.midi added (display for incoming messages)
   - ui.vkeyboard added (with ui.vk alias for vertical keyboard)